Refactor component & enhance ai usage & add Docs Fixes #1
