+++ .env
@@ -1,1 +1,2 @@
 VITE_API_KEY=01b35711f29f4fa98738baa18660c9bf
+VITE_MIMETYPE=audio/webm
