+++ src/components/AudioRecorder.tsx
@@ -1,78 +1,10 @@
 import { useState, useRef, useEffect } from 'react';
-import axios from 'axios';
 import Status from './Status';
 import Result from './Result';
+import { Transcript } from '../types/transcript';
+import { assemblyAPI } from '../core/assemblyAPI';
 
-const assemblyAPI = axios.create({
-  baseURL: 'https://api.assemblyai.com/v2',
-  headers: {
-    authorization: process.env.VITE_API_KEY,
-    'content-type': 'application/json',
-  },
-});
-
-const mimeType = 'audio/webm';
-
-export interface Entities {
-  entity_type: string;
-  text: string;
-  start?: number;
-  end?: number;
-}
-export interface Words {
-  text: string;
-  start: number;
-  end: number;
-  confidence: number;
-  speaker: string;
-}
-export interface Utterance {
-  confidence: number;
-  end: number;
-  speaker: string;
-  start: number;
-  text: string;
-  words: Words[];
-}
-export interface SentimentAnalysisResults {
-  confidence: number;
-  text: string;
-  sentiment: 'POSITIVE' | 'NEGATIVE' | 'NEUTRAL';
-}
-
-export interface Label {
-  relevance: number;
-  label: string;
-}
-
-export interface Result {
-  text: string;
-  labels: Label[];
-  timestamp: {
-    start: number;
-    end: number;
-  };
-}
-
-export interface Summary {
-  topic: number;
-}
-
-export interface IabCategoriesResult {
-  status: 'success' | 'unavailable';
-  results: Result[];
-  summary: Summary;
-}
-
-export interface Transcript {
-  id: string;
-  status?: 'processing' | 'queued' | 'completed' | 'error';
-  text?: string;
-  sentiment_analysis_results?: SentimentAnalysisResults[];
-  entities?: Entities[];
-  utterances?: Utterance[];
-  iab_categories_result?: IabCategoriesResult;
-}
+const mimeType = process.env.VITE_MIMETYPE;
 
 const AudioRecorder = () => {
   const [permission, setPermission] = useState(false);
