+++ src/core/assemblyAPI.ts
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+export const assemblyAPI = axios.create({
+  baseURL: 'https://api.assemblyai.com/v2',
+  headers: {
+    authorization: process.env.VITE_API_KEY,
+    'content-type': 'application/json',
+  },
+});
